requirements.txt
playwright
pytest
pytest-playwright
pytest-html
requests
python-dotenv



